requirements.txt
django==6.0.3
redis==7.2.1
requests==2.32.5
pillow==12.1.1
django-bootstrap5==26.2
django-debug-toolbar==6.2.0
